move preloadTiles to separate file
